--- a/docker/docker-cluster/metrictank.ini
+++ b/docker/docker-cluster/metrictank.ini
@@ -291,3 +291,7 @@ create-keyspace = false
 ### in-memory only
 [memory-idx]
 enabled = false
+# enables/disables querying based on tags
+tag-support = false
+# size of regular expression cache in tag query evaluation
+match-cache-size = 1000