--- a/docker/docker-dev-custom-cfg-kafka/metrictank.ini
+++ b/docker/docker-dev-custom-cfg-kafka/metrictank.ini
@@ -291,3 +291,7 @@ create-keyspace = true
 ### in-memory only
 [memory-idx]
 enabled = false
+# enables/disables querying based on tags
+tag-support = false
+# size of regular expression cache in tag query evaluation
+match-cache-size = 1000