--- a/idx/idx.go
+++ b/idx/idx.go
@@ -101,7 +101,25 @@ Interface
   passed is -1, then the all orgs should be examined for stale metrics to be deleted.
   The method returns a list of the metricDefinitions deleted from the index and any
   error encountered.
+
+* TagList(int) []string:
+  Returns a list of all tag keys associated with the metrics of a given organization.o
+
+* Tag(int, string, int64) map[string]uint32:
+  Returns a list of all values associated with a given tag key in the given org. The
+  occurences of each value is counted and the count is referred to by the series ids
+  in the returned map. If the third parameter is > 0 then the metrics will be filtered
+  and only those of which the LastUpdate time is >= the from timestamp will be
+  considered while the others are being ignored.
+
+* IdsByTagExpression(int, []string, int64) ([]string, error):
+  Takes a list of expressions in the format key=value.
+  The allowed operators are: =, !=, =~, !=~.
+  Returns a slice of IDs that match the given conditions, the conditions are
+  logically AND-ed. If the third argument is > 0 then the results will be filtered
+  and only those where the LastUpdate time is >= from will be returned as results.
 */
+
 type MetricIndex interface {
 	Init() error
 	Stop()