--- a/idx/idx.go
+++ b/idx/idx.go
@@ -113,6 +113,6 @@ type MetricIndex interface {
 	List(int) []Archive
 	Prune(int, time.Time) ([]Archive, error)
 	TagList(int) []string
-	Tag(int, string) map[string]uint32
-	IdsByTagExpressions(int, []string) ([]string, error)
+	Tag(int, string, int64) map[string]uint32
+	IdsByTagExpressions(int, []string, int64) ([]string, error)
 }