--- a/idx/idx.go
+++ b/idx/idx.go
@@ -155,7 +155,7 @@ Interface
   will be filtered and only those of which the LastUpdate time is >= the from
   timestamp will be considered while the others are being ignored.
 
-* IdsByTagExpression(int, []string, int64) ([]string, error):
+* FindByTag(int, []string, int64) ([]string, error):
   This method takes a list of expressions in the format key<operator>value.
   The allowed operators are: =, !=, =~, !=~.
   It returns a slice of IDs that match the given conditions, the conditions are
@@ -175,5 +175,5 @@ type MetricIndex interface {
 	Prune(int, time.Time) ([]Archive, error)
 	TagList(int) []string
 	Tag(int, string, int64) map[string]uint32
-	IdsByTagExpressions(int, []string, int64) (map[MetricID]struct{}, error)
+	FindByTag(int, []string, int64) (map[MetricID]struct{}, error)
 }