--- a/idx/memory/memory.go
+++ b/idx/memory/memory.go
@@ -51,6 +51,7 @@ func ConfigSetup() {
 	memoryIdx := flag.NewFlagSet("memory-idx", flag.ExitOnError)
 	memoryIdx.BoolVar(&Enabled, "enabled", false, "")
 	memoryIdx.BoolVar(&tagSupport, "tag-support", false, "")
+	memoryIdx.IntVar(&matchCacheSize, "match-cache-size", 1000, "")
 	globalconf.Register("memory-idx", memoryIdx)
 }