--- a/idx/memory/memory.go
+++ b/idx/memory/memory.go
@@ -454,8 +454,7 @@ func (m *MemoryIdx) FindByTag(orgId int, expressions []string, from int64) (map[
 		return nil, err
 	}
 
-	seriesMap := m.idsByTagQuery(orgId, query)
-	return seriesMap, nil
+	return m.idsByTagQuery(orgId, query), nil
 }
 
 func (m *MemoryIdx) idsByTagQuery(orgId int, query TagQuery) TagIDs {
@@ -467,9 +466,7 @@ func (m *MemoryIdx) idsByTagQuery(orgId int, query TagQuery) TagIDs {
 		return nil
 	}
 
-	res := query.Run(tree, m.DefById)
-
-	return res
+	return query.Run(tree, m.DefById)
 }
 
 func (m *MemoryIdx) Find(orgId int, pattern string, from int64) ([]idx.Node, error) {