--- a/metrictank-sample.ini
+++ b/metrictank-sample.ini
@@ -294,3 +294,7 @@ create-keyspace = true
 ### in-memory only
 [memory-idx]
 enabled = false
+# enables/disables querying based on tags
+tag-support = false
+# size of regular expression cache in tag query evaluation
+match-cache-size = 1000