@Dieterbe 2017-09-30
FTR: prior discussion: #532 
feel free to link to any other tickets or discussions