@replay 2017-10-04
Yesterday @DanCech and I came up with some new ideas how to optimize the tag intersecting/filtering when the query contains a large number of expressions. 
The basic idea is that we're trying to be smart about selecting the expression that is going to result in the smallest result set and is the fastest to execute, once we got the initial result set from that expression we simply use all the other expressions to filter ids out of the result set in increasing order by how cpu-heavy they are. This removes the need to intersect multiple result sets and only uses filtering of a list of potential results.

This is the patch that implements this change: https://github.com/grafana/metrictank/commit/5d94fe39c2d498502b2c915d52d305c43d4419a2
I added a benchmark specifically to compare the performance before and after this change: https://github.com/grafana/metrictank/commit/d7ff88d7b71a423ac2e4463606c2ea8558614428

That's the results I get on my laptop:
```
go test github.com/grafana/metrictank/idx/memory  -test.v   -bench=BenchmarkTagQueryFilterAndIntersect    -run=BenchmarkTagQueryFilterAndIntersect -benchtime=10s 

--- before the optimization ---
goos: linux
goarch: amd64
pkg: github.com/grafana/metrictank/idx/memory
BenchmarkTagQueryFilterAndIntersect-8   	     100	 286923713 ns/op	190802678 B/op	  484214 allocs/op
PASS
ok  	github.com/grafana/metrictank/idx/memory	66.783s

--- after the optimization ---
goos: linux
goarch: amd64
pkg: github.com/grafana/metrictank/idx/memory
BenchmarkTagQueryFilterAndIntersect-8   	     100	 108550842 ns/op	 9124985 B/op	  111516 allocs/op
PASS
ok  	github.com/grafana/metrictank/idx/memory	46.103s
```