@Dieterbe 2017-10-09
I did the same bench, but using docker-dev, because single node means i can pump in more data using less RAM, and bypassing kafka io constraints, and it removes noise from clustering. also it doesn't need cache warmup since it keeps 60~70min worth of data in ring buffers.
first:
```
fakemetrics -carbon-tcp-address localhost:2003 -statsd-addr localhost:8125 -shard-org -keys-per-org 30000 -orgs 1 -speedup 5 -stop-at-now -offset 60min; echo realtime; fakemetrics -carbon-tcp-address localhost:2003 -statsd-addr localhost:8125 -shard-org -keys-per-org 30000 -orgs 1
```

## master


https://snapshot.raintank.io/dashboard/snapshot/vyKVx2LSep7jmsC0kkcaAZXLrIv6RRoQ
```
./mt-index-cat -prefix='some.id.of.a.metric.1111' -from=1h -addr http://localhost:6060 cass localhost:9042 metrictank vegeta-render | vegeta attack -rate 100 -duration=2m | vegeta report
Requests      [total, rate]            12000, 100.01
Duration      [total, attack, wait]    1m59.992730822s, 1m59.98999992s, 2.730902ms
Latencies     [mean, 50, 95, 99, max]  3.425299ms, 2.962018ms, 5.321823ms, 16.326446ms, 46.338665ms
Bytes In      [total, mean]            189946597, 15828.88
Bytes Out     [total, mean]            0, 0.00
Success       [ratio]                  100.00%
Status Codes  [code:count]             200:12000  
Error Set:
```

## this PR

https://snapshot.raintank.io/dashboard/snapshot/JQBTO549Vl8vqY97R70k830dK6yAB68O?orgId=2

```
./mt-index-cat -prefix='some.id.of.a.metric.1111' -from=1h -addr http://localhost:6060 cass localhost:9042 metrictank vegeta-render | vegeta attack -rate 100 -duration=2m | vegeta report
Requests      [total, rate]            12000, 100.01
Duration      [total, attack, wait]    1m59.99311947s, 1m59.989999929s, 3.119541ms
Latencies     [mean, 50, 95, 99, max]  4.228529ms, 3.16475ms, 14.868404ms, 16.881767ms, 64.243791ms
Bytes In      [total, mean]            189947129, 15828.93
Bytes Out     [total, mean]            0, 0.00
Success       [ratio]                  100.00%
Status Codes  [code:count]             200:12000  
Error Set:

```

## conclusion

* seems similar as your first benchmark. comparing the mean shows a 23% latency increase. the median is only 6.8% slower so it's mostly due to the outliers.
* note that the dashboards also show number of heap objects. and they look about the same, in fact, it semes slightly lower for tag-index case.
* also GC cpu fraction is lowered in tag-index case, but that might be misleading
* given the ~ of heap objects, and that I only tested once, and you tested twice and came out non-conclusive, i'm not super confident in any of our benchmarks at this point. any idea why your 2nd benchmark was so different?
* either way, for now it looks like performance is not a stopper. we can go forward and try to optimize more down the road (GC or otherwise)